convert Utilities.js to ts, add type predicate logic
